Move classes according to review feedback
