+++ src/ext-test/java/org/opentripplanner/ext/restapi/parameter/ApiRequestModeTest.java
@@ -1,10 +1,12 @@
-package org.opentripplanner.api.parameter;
+package org.opentripplanner.ext.restapi.parameter;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.opentripplanner.transit.model.basic.TransitMode.CARPOOL;
 
 import java.util.List;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
+import org.opentripplanner.api.parameter.ApiRequestMode;
 
 class ApiRequestModeTest {
