+++ src/ext-test/java/org/opentripplanner/ext/restapi/parameter/QualifiedModeTest.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.api.parameter;
+package org.opentripplanner.ext.restapi.parameter;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -6,6 +6,9 @@ import java.util.Arrays;
 import java.util.Set;
 import java.util.stream.Collectors;
 import org.junit.jupiter.api.Test;
+import org.opentripplanner.api.parameter.ApiRequestMode;
+import org.opentripplanner.api.parameter.QualifiedMode;
+import org.opentripplanner.api.parameter.Qualifier;
 
 public class QualifiedModeTest {
