+++ src/ext/java/org/opentripplanner/ext/restapi/resources/PlannerResource.java
@@ -11,7 +11,7 @@ import jakarta.ws.rs.core.UriInfo;
 import org.glassfish.grizzly.http.server.Request;
 import org.opentripplanner.api.common.Message;
 import org.opentripplanner.api.model.error.PlannerError;
-import org.opentripplanner.apis.common.mapping.PlannerErrorMapper;
+import org.opentripplanner.apis.support.mapping.PlannerErrorMapper;
 import org.opentripplanner.ext.restapi.mapping.TripPlanMapper;
 import org.opentripplanner.ext.restapi.mapping.TripSearchMetadataMapper;
 import org.opentripplanner.ext.restapi.model.ElevationMetadata;
