+++ src/ext/java/org/opentripplanner/ext/vectortiles/layers/vehicleparkings/StadtnaviVehicleParkingPropertyMapper.java
@@ -4,7 +4,7 @@ import java.util.Collection;
 import java.util.List;
 import java.util.Locale;
 import org.json.simple.JSONObject;
-import org.opentripplanner.apis.common.mapping.PropertyMapper;
+import org.opentripplanner.apis.support.mapping.PropertyMapper;
 import org.opentripplanner.framework.i18n.I18NStringMapper;
 import org.opentripplanner.inspector.vector.KeyValue;
 import org.opentripplanner.model.calendar.openinghours.OsmOpeningHoursSupport;
