+++ src/ext/java/org/opentripplanner/ext/vectortiles/layers/vehiclerental/mapper/DigitransitVehicleRentalPropertyMapper.java
@@ -3,7 +3,7 @@ package org.opentripplanner.ext.vectortiles.layers.vehiclerental.mapper;
 import java.util.Collection;
 import java.util.List;
 import java.util.Locale;
-import org.opentripplanner.apis.common.mapping.PropertyMapper;
+import org.opentripplanner.apis.support.mapping.PropertyMapper;
 import org.opentripplanner.framework.i18n.I18NStringMapper;
 import org.opentripplanner.inspector.vector.KeyValue;
 import org.opentripplanner.service.vehiclerental.model.VehicleRentalPlace;
