+++ src/main/java/org/opentripplanner/apis/gtfs/datafetchers/PlanImpl.java
@@ -4,8 +4,8 @@ import graphql.schema.DataFetcher;
 import graphql.schema.DataFetchingEnvironment;
 import java.util.stream.Collectors;
 import org.opentripplanner.api.resource.DebugOutput;
-import org.opentripplanner.apis.common.mapping.PlannerErrorMapper;
 import org.opentripplanner.apis.gtfs.generated.GraphQLDataFetchers;
+import org.opentripplanner.apis.support.mapping.PlannerErrorMapper;
 import org.opentripplanner.model.plan.Itinerary;
 import org.opentripplanner.model.plan.StopArrival;
 import org.opentripplanner.model.plan.paging.cursor.PageCursor;
