+++ src/main/java/org/opentripplanner/inspector/vector/AreaStopsLayerBuilder.java
@@ -7,7 +7,7 @@ import java.util.Map;
 import java.util.function.Function;
 import org.locationtech.jts.geom.Envelope;
 import org.locationtech.jts.geom.Geometry;
-import org.opentripplanner.apis.common.mapping.PropertyMapper;
+import org.opentripplanner.apis.support.mapping.PropertyMapper;
 import org.opentripplanner.transit.model.site.AreaStop;
 import org.opentripplanner.transit.service.TransitService;
