+++ src/main/java/org/opentripplanner/inspector/vector/DebugClientAreaStopPropertyMapper.java
@@ -3,7 +3,7 @@ package org.opentripplanner.inspector.vector;
 import java.util.Collection;
 import java.util.List;
 import java.util.Locale;
-import org.opentripplanner.apis.common.mapping.PropertyMapper;
+import org.opentripplanner.apis.support.mapping.PropertyMapper;
 import org.opentripplanner.framework.i18n.I18NStringMapper;
 import org.opentripplanner.transit.model.site.AreaStop;
 import org.opentripplanner.transit.service.TransitService;
